Split scribe.py: extract caption.py + summary.py (+ DocUnit -> parse.py)
Phase-0 refactor (#306), commit 4 of 5. Behaviour-preserving.

- bartleby/ingest/caption.py — phase 2: _analyze_image, _caption_from_analysis,
  _caption_all (the decoupled, concurrent image-captioning pass, #166).
- bartleby/ingest/summary.py — phase 3: _summary_chunks, _summarize_all (the
  concurrent summarize pass, #167/#188), kept separate from summarize.py to
  avoid the writer<->summarize import cycle. _summary_chunks routes through
  embed.embed_texts and parsers._build_chunk_inputs.
- bartleby/ingest/parse.py — seeded with DocUnit (renamed from _DocUnit), the
  per-document unit caption/summary annotate; parse_all() lands in commit 5.

scribe.py imports caption/summary/parse and calls caption._caption_all /
summary._summarize_all / parse.DocUnit; the orphaned imports (itertools,
threading, Callable, concurrent.futures, dataclass, ChunkInput, image_pipeline,
chunk_markdown_string, embed_texts, summarize, the image/summary writer types,
_ProgressTally, Provider) are dropped. Tests retargeted; the dead
scribe.embed_texts patch removed from mock_embed (embed.embed_texts covers it).
