+++ bartleby/commands/scribe.py
@@ -26,15 +26,10 @@ documents are unwrapped and routed individually.
 
 from __future__ import annotations
 
-import itertools
 import json
 import sys
-import threading
 import time
-from concurrent.futures import FIRST_COMPLETED, ThreadPoolExecutor, as_completed, wait
-from dataclasses import dataclass
 from pathlib import Path
-from typing import Callable
 
 from loguru import logger
 
@@ -43,27 +38,17 @@ from bartleby.config import (
     load_config,
     redact_config,
 )
-from bartleby.db.chunks import ChunkInput
 from bartleby.db.connection import open_db, resolve_project_name
+from bartleby.ingest import caption
 from bartleby.ingest import classify
-from bartleby.ingest import images as image_pipeline
+from bartleby.ingest import parse
 from bartleby.ingest import parsers
 from bartleby.ingest import pool
 from bartleby.ingest import resolve
-from bartleby.ingest.progress import ScribeProgress, _ProgressTally
-from bartleby.ingest.chunk import (
-    chunk_markdown_string,
-    resolve_format_filter,
-)
-from bartleby.ingest.embed import embed_texts
-from bartleby.ingest.summarize import summarize
-from bartleby.ingest.writer import (
-    MAX_INGEST_ATTEMPTS,
-    ImageCaption,
-    PendingImage,
-    PendingSummary,
-    Writer,
-)
+from bartleby.ingest import summary
+from bartleby.ingest.progress import ScribeProgress
+from bartleby.ingest.chunk import resolve_format_filter
+from bartleby.ingest.writer import MAX_INGEST_ATTEMPTS, Writer
 from bartleby.lib import console
 from bartleby.lib import timing
 from bartleby.lib.consts import (
@@ -75,310 +60,6 @@ from bartleby.lib.consts import (
     DEFAULT_VISION_MIN_DIMENSION,
 )
 from bartleby.project import get_project_dir
-from bartleby.providers import Provider
-
-
-# -------------------- shared helpers --------------------
-
-
-def _summary_chunks(text: str) -> list[ChunkInput]:
-    """Chunk + embed a summary's markdown body into ChunkInputs (producer side)."""
-    rows = chunk_markdown_string(text)
-    if not rows:
-        return []
-    embeddings = embed_texts([r.text for r in rows])
-    return parsers._build_chunk_inputs(rows, embeddings)
-
-
-# -------------------- image caption helpers --------------------
-
-
-def _analyze_image(
-    pending: PendingImage,
-    *,
-    vision_provider: Provider,
-    vision_model: str,
-) -> image_pipeline.ImageAnalysis:
-    """Run the §7 image pipeline (Tesseract → VLM) on one recorded image row.
-
-    Reloads the prepared JPEG from the archive and classifies text-image vs
-    scene-image (VLM only for scenes). Pure analysis: no embedding, no DB — the
-    OCR subprocess and VLM network call both release the GIL, so this is the
-    half of captioning that runs off the writer thread in the caption pool.
-    """
-    prepared = image_pipeline.PreparedImage(
-        hash=pending.file_hash,
-        jpeg_bytes=Path(pending.file_path).read_bytes(),
-        width=pending.width,
-        height=pending.height,
-    )
-    return image_pipeline.analyze(vision_provider, prepared, model=vision_model)
-
-
-def _caption_from_analysis(
-    pending: PendingImage,
-    analysis: image_pipeline.ImageAnalysis,
-    vision_model: str,
-) -> ImageCaption:
-    """Package an analysis into a writer-ready caption, embedding its chunk.
-
-    Runs on the writer thread: ``analysis_to_chunk_inputs`` embeds via the one
-    per-process SentenceTransformer, which isn't safe to call from several
-    caption threads at once — so embedding stays here, next to the DB write.
-    """
-    return ImageCaption(
-        image_id=pending.image_id,
-        analysis_json=analysis.model_dump_json(),
-        analysis_model=vision_model,
-        chunks=image_pipeline.analysis_to_chunk_inputs(analysis),
-    )
-
-
-# -------------------- per-file orchestration --------------------
-
-
-@dataclass
-class _DocUnit:
-    """A document parsed + persisted this run (or resumed from an earlier one),
-    carried through the caption and summarize phases. ``stages`` accumulates
-    per-stage seconds only under ``--timings`` — it's None on the fast path."""
-    document_id: int
-    file_name: str
-    file_hash: str
-    page_count: int | None = None
-    stages: dict[str, float] | None = None
-
-
-def _caption_all(
-    writer: Writer,
-    units: list[_DocUnit],
-    *,
-    vision_provider: Provider | None,
-    vision_model: str | None,
-    vision_enabled: bool,
-    caption_workers: int,
-    timings: bool,
-    on_progress: Callable[[int, int], None] | None,
-    on_lane: Callable[[object, str, str], None] | None = None,
-) -> None:
-    """Phase 2: caption every still-uncaptioned image across all parsed units.
-
-    Captioning is decoupled from parse (#166): parse only records image rows
-    (deduped by byte-hash, ``analysis_json IS NULL``); this stage analyzes them
-    and joins the captions back through the single Writer. Rows are keyed by id,
-    so one image shared by several documents is captioned once. Idempotent: only
-    rows still lacking a caption are touched, so a resumed run fills gaps.
-
-    Analysis (OCR + VLM) runs concurrently across ``caption_workers`` threads —
-    the network-bound work the GIL releases — while embedding and the DB write
-    stay on this thread, the Writer's sole owner. At ``caption_workers <= 1`` the
-    work runs inline: the path ``--timings`` forces, for a clean sequential
-    baseline with per-document caption seconds.
-
-    Mutates nothing on the units except their ``stages`` (under timings); each
-    document's remaining incompleteness is recomputed from the DB in phase 3.
-    """
-    # Unique uncaptioned rows across every unit; the first unit referencing a row
-    # owns its timing. A row shared across documents appears once.
-    pending: dict[int, PendingImage] = {}
-    owner: dict[int, _DocUnit] = {}
-    for unit in units:
-        for pi in writer.uncaptioned_images(unit.document_id):
-            if pi.image_id not in pending:
-                pending[pi.image_id] = pi
-                owner[pi.image_id] = unit
-    if not pending:
-        return
-    if not vision_enabled:
-        console.warn(
-            f"Skipping {len(pending)} uncaptioned image(s) — no vision provider "
-            f"configured."
-        )
-        return
-
-    progress = _ProgressTally(len(pending), on_progress)
-
-    # Capped rows (failed MAX_INGEST_ATTEMPTS× already) are skipped, not retried.
-    to_caption: dict[int, PendingImage] = {}
-    for image_id, pi in pending.items():
-        if writer.is_capped(pi.file_hash, "caption"):
-            console.warn(
-                f"{owner[image_id].file_name}: skipping image — failed "
-                f"{MAX_INGEST_ATTEMPTS}× already; not retrying."
-            )
-            progress.advance()
-        else:
-            to_caption[image_id] = pi
-
-    def _persist(image_id: int, analysis: image_pipeline.ImageAnalysis) -> None:
-        pi = to_caption[image_id]
-        writer.persist_caption(_caption_from_analysis(pi, analysis, vision_model))
-        writer.clear_failure(pi.file_hash, "caption")
-
-    def _fail(image_id: int, exc: Exception) -> None:
-        unit = owner[image_id]
-        console.warn(f"{unit.file_name}: image analysis failed: {exc}")
-        writer.record_failure(
-            to_caption[image_id].file_hash, unit.file_name, "caption", exc
-        )
-
-    def _analyze(image_id: int, pi: PendingImage) -> image_pipeline.ImageAnalysis:
-        # Claim this thread's lane for the owning document, then run the analysis.
-        # Keyed by thread id, so the pool's N threads map to N sticky lanes.
-        if on_lane is not None:
-            on_lane(threading.get_ident(), owner[image_id].file_name, "captioning")
-        return _analyze_image(
-            pi, vision_provider=vision_provider, vision_model=vision_model,
-        )
-
-    if caption_workers <= 1:
-        # Inline: one image at a time on this thread. Same analyze/persist calls
-        # as the pool, but with clean per-document caption seconds for --timings.
-        for image_id, pi in to_caption.items():
-            t0 = time.perf_counter() if timings else None
-            try:
-                _persist(image_id, _analyze(image_id, pi))
-            except Exception as e:
-                _fail(image_id, e)
-            unit = owner[image_id]
-            if t0 is not None and unit.stages is not None:
-                unit.stages["caption"] = (
-                    unit.stages.get("caption", 0.0) + (time.perf_counter() - t0)
-                )
-            progress.advance()
-        return
-
-    with ThreadPoolExecutor(max_workers=caption_workers) as pool:
-        futures = {
-            pool.submit(_analyze, image_id, pi): image_id
-            for image_id, pi in to_caption.items()
-        }
-        for fut in as_completed(futures):
-            image_id = futures[fut]
-            try:
-                _persist(image_id, fut.result())
-            except Exception as e:
-                _fail(image_id, e)
-            progress.advance()
-
-
-def _summarize_all(
-    writer: Writer,
-    pending: list[PendingSummary],
-    *,
-    llm_provider: Provider | None,
-    llm_model: str | None,
-    temperature: float,
-    max_summarize_tokens: int,
-    summarize_workers: int,
-    timings: bool,
-    on_progress: Callable[[int, int], None] | None,
-    on_lane: Callable[[object, str, str], None] | None = None,
-    reasoning_effort: str | None = None,
-) -> tuple[int, dict[int, tuple[str, float]]]:
-    """Phase 3: summarize every document still owing one, concurrently (#188).
-
-    Mirrors the caption stage (#166): the network-bound ``summarize()`` LLM call
-    fans out across ``summarize_workers`` threads while the Writer's persist stays
-    on this thread, its connection's sole owner. Capped documents are surfaced as
-    still-incomplete rather than retried; a single failure never tears down the
-    pass. The ``pending`` work-list is :meth:`Writer.documents_needing_summary`,
-    which already excludes summarized and zero-chunk docs (issue #80), so the only
-    main-thread guard left is the cap check.
-
-    One deliberate difference from captioning: a caption's payload is an on-disk
-    image the worker reads itself, but a summary's payload is assembled from the
-    DB by ``writer.summary_input`` — a Writer-owned read that must stay on this
-    thread. So inputs are fetched here and only ``summarize_workers`` are kept in
-    flight, holding the old sequential pass's flat memory footprint instead of
-    materializing every document's text at once.
-
-    Returns ``(incomplete_count, {document_id: (file_name, seconds)})`` — the
-    second map is populated only under ``timings`` (which forces one worker, so
-    the per-document seconds stay a clean sequential baseline).
-    """
-    incomplete = 0
-    times: dict[int, tuple[str, float]] = {}
-    progress = _ProgressTally(len(pending), on_progress)
-
-    # Capped docs (failed MAX_INGEST_ATTEMPTS× already) are surfaced, not retried.
-    work: list[PendingSummary] = []
-    for ps in pending:
-        if writer.is_capped(ps.file_hash, "summary"):
-            console.warn(
-                f"{ps.file_name}: skipping summary — failed "
-                f"{MAX_INGEST_ATTEMPTS}× already; not retrying."
-            )
-            incomplete += 1
-            progress.advance()
-        else:
-            work.append(ps)
-
-    def _persist(ps: PendingSummary, result) -> None:
-        writer.persist_summary(
-            ps.document_id, result, _summary_chunks(result.text)
-        )
-        writer.clear_failure(ps.file_hash, "summary")
-
-    def _fail(ps: PendingSummary, exc: Exception) -> None:
-        nonlocal incomplete
-        console.warn(f"{ps.file_name}: summary failed: {exc}")
-        writer.record_failure(ps.file_hash, ps.file_name, "summary", exc)
-        incomplete += 1
-
-    def _summarize(ps: PendingSummary, text: str):
-        # Claim the running thread's lane for this document, then make the LLM
-        # call — keyed by thread id so the pool's N threads map to N sticky lanes.
-        if on_lane is not None:
-            on_lane(threading.get_ident(), ps.file_name, "summarizing")
-        return summarize(
-            text, provider=llm_provider, model=llm_model,
-            temperature=temperature, max_summarize_tokens=max_summarize_tokens,
-            reasoning_effort=reasoning_effort,
-        )
-
-    if summarize_workers <= 1:
-        # Inline: one document at a time on this thread. Same summarize/persist
-        # calls as the pool, but with clean per-document seconds for --timings.
-        for ps in work:
-            t0 = time.perf_counter() if timings else None
-            try:
-                _persist(ps, _summarize(ps, writer.summary_input(ps.document_id)))
-            except Exception as e:
-                _fail(ps, e)
-            if t0 is not None:
-                times[ps.document_id] = (ps.file_name, time.perf_counter() - t0)
-            progress.advance()
-        return incomplete, times
-
-    # Pooled: keep at most ``summarize_workers`` inputs in flight (see docstring) —
-    # fetch each document's text on this thread, then top up one as each completes.
-    it = iter(work)
-    with ThreadPoolExecutor(max_workers=summarize_workers) as pool:
-        in_flight = {
-            pool.submit(_summarize, ps, writer.summary_input(ps.document_id)): ps
-            for ps in itertools.islice(it, summarize_workers)
-        }
-        while in_flight:
-            finished, _ = wait(in_flight, return_when=FIRST_COMPLETED)
-            for fut in finished:
-                ps = in_flight.pop(fut)
-                try:
-                    _persist(ps, fut.result())
-                except Exception as e:
-                    _fail(ps, e)
-                progress.advance()
-                nxt = next(it, None)
-                if nxt is not None:
-                    in_flight[
-                        pool.submit(
-                            _summarize, nxt, writer.summary_input(nxt.document_id)
-                        )
-                    ] = nxt
-    return incomplete, times
-
-
-# -------------------- entry --------------------
 
 
 def _report_failures(writer: Writer) -> None:
@@ -590,8 +271,8 @@ def main(
             # their own pass over the DB (#167). Under --timings each unit's record
             # starts from parse and gains caption/summarize downstream.
             incomplete_count = 0
-            units: list[_DocUnit] = [
-                _DocUnit(
+            units: list[parse.DocUnit] = [
+                parse.DocUnit(
                     item.document_id, item.file_name, item.file_hash,
                     stages={} if timings else None,
                 )
@@ -640,7 +321,7 @@ def main(
                     stages["embed"] = (
                         stages.get("embed", 0.0) + (time.perf_counter() - persist_t)
                     )
-                units.append(_DocUnit(
+                units.append(parse.DocUnit(
                     document_id, req.file_name, req.file_hash,
                     page_count=outcome.parsed.page_count, stages=stages,
                 ))
@@ -652,7 +333,7 @@ def main(
             # adapts the tally contract itself and reveals the caption total the
             # first time it hears one.
             cap_phase = progress.phase("caption")
-            _caption_all(
+            caption._caption_all(
                 writer, units,
                 vision_provider=vision_provider, vision_model=vision_model,
                 vision_enabled=parse_config.vision_enabled,
@@ -686,7 +367,7 @@ def main(
                 pending = writer.documents_needing_summary()
                 if pending:
                     sum_phase = progress.phase("summarize")
-                    owed, summarize_times = _summarize_all(
+                    owed, summarize_times = summary._summarize_all(
                         writer, pending,
                         llm_provider=llm_provider, llm_model=llm_model,
                         temperature=temperature,
